Atividade_BD/requirements.txt
bcrypt==4.3.0
blinker==1.9.0
click==8.2.1
colorama==0.4.6
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
psycopg==3.2.9
psycopg-binary==3.2.9
psycopg-pool==3.2.6
python-dotenv==1.0.1
typing_extensions==4.14.1
tzdata==2025.2
Werkzeug==3.1.3
